Move survey selector to features
