+++ packages/datatrak-web/src/features/index.ts
@@ -22,3 +22,4 @@ export { Leaderboard } from './Leaderboard';
 export { Reports } from './Reports';
 export { TaskPageHeader, TasksTable, CreateTaskModal } from './Tasks';
 export { useUserCountries, CountrySelector } from './CountrySelector';
+export { GroupedSurveyList } from './GroupedSurveyList';
