+++ packages/datatrak-web/src/views/SurveySelectPage.tsx
@@ -8,11 +8,10 @@ import styled from 'styled-components';
 import { DialogActions, Paper, Typography } from '@material-ui/core';
 import { SpinningLoader } from '@tupaia/ui-components';
 import { useEditUser } from '../api/mutations';
-import { SelectList, ListItemType, Button, SurveyFolderIcon, SurveyIcon } from '../components';
-import { Survey } from '../types';
+import { ListItemType, Button } from '../components';
 import { useCurrentUserContext, useProjectSurveys } from '../api';
 import { HEADER_HEIGHT } from '../constants';
-import { CountrySelector, useUserCountries } from '../features';
+import { CountrySelector, GroupedSurveyList, useUserCountries } from '../features';
 
 const Container = styled(Paper).attrs({
   variant: 'outlined',
@@ -51,17 +50,6 @@ const LoadingContainer = styled.div`
   flex: 1;
 `;
 
-const ListWrapper = styled.div`
-  max-height: 35rem;
-  display: flex;
-  flex-direction: column;
-  overflow: auto;
-  flex: 1;
-  ${({ theme }) => theme.breakpoints.down('sm')} {
-    max-height: 100%;
-  }
-`;
-
 const HeaderWrapper = styled.div`
   display: flex;
   align-items: center;
@@ -92,12 +80,6 @@ const Subheader = styled(Typography).attrs({
   }
 `;
 
-const sortAlphanumerically = (a: ListItemType, b: ListItemType) => {
-  return (a.content as string).trim()?.localeCompare((b.content as string).trim(), 'en', {
-    numeric: true,
-  });
-};
-
 export const SurveySelectPage = () => {
   const navigate = useNavigate();
   const [selectedSurvey, setSelectedSurvey] = useState<ListItemType | null>(null);
@@ -115,48 +97,6 @@ export const SurveySelectPage = () => {
   const user = useCurrentUserContext();
 
   const { data: surveys, isLoading } = useProjectSurveys(user.projectId, selectedCountry?.name);
-
-  // group the data by surveyGroupName for the list, and add the value and selected properties
-  const groupedSurveys =
-    surveys
-      ?.reduce((acc: ListItemType[], survey: Survey) => {
-        const { surveyGroupName, name, code } = survey;
-        const formattedSurvey = {
-          content: name,
-          value: code,
-          selected: selectedSurvey?.value === code,
-          icon: <SurveyIcon />,
-        };
-        // if there is no surveyGroupName, add the survey to the list as a top level item
-        if (!surveyGroupName) {
-          return [...acc, formattedSurvey];
-        }
-        const group = acc.find(({ content }) => content === surveyGroupName);
-        // if the surveyGroupName doesn't exist in the list, add it as a top level item
-        if (!group) {
-          return [
-            ...acc,
-            {
-              content: surveyGroupName,
-              icon: <SurveyFolderIcon />,
-              value: surveyGroupName,
-              children: [formattedSurvey],
-            },
-          ];
-        }
-        // if the surveyGroupName exists in the list, add the survey to the children
-        return acc.map(item => {
-          if (item.content === surveyGroupName) {
-            return {
-              ...item,
-              // sort the folder items alphanumerically
-              children: [...(item.children || []), formattedSurvey].sort(sortAlphanumerically),
-            };
-          }
-          return item;
-        });
-      }, [])
-      ?.sort(sortAlphanumerically) ?? [];
 
   const handleSelectSurvey = () => {
     if (countryHasUpdated) {
@@ -191,9 +131,11 @@ export const SurveySelectPage = () => {
           <SpinningLoader />
         </LoadingContainer>
       ) : (
-        <ListWrapper>
-          <SelectList items={groupedSurveys} onSelect={setSelectedSurvey} />
-        </ListWrapper>
+        <GroupedSurveyList
+          surveys={surveys}
+          setSelectedSurvey={setSelectedSurvey}
+          selectedSurvey={selectedSurvey}
+        />
       )}
       <DialogActions>
         <Button to="/" variant="outlined">
